TicketTerminal API v17

+++ src/main/java/com/ticket/terminal/dto/VisitObjectItemDto.java
@@ -28,7 +28,7 @@ public class VisitObjectItemDto {
     private Long groupVisitObjectId;
 
     @JsonProperty("CategoryVisitorId")
-    private Integer categoryVisitorId;
+    private Long categoryVisitorId;
 
     @JsonProperty("Address")
     private String address;
